External glusterfs as PV
Currently for each existing gluster volume, one needs to provide
a different storage class, as specified in the example file.

In future we can extend this to also create newer glusterfs subvols
to provide PVs.

This is just to validate our concept that if there is a PV provisioned
in heketi based deployment, and one can use the same glusterfs volume
with kadalu in later k8s version.

Also added tests

Updates: #36