working external storage
How this works:

Gluster Host, and volume name are passed to kadalu through
storage class parameters, with a keyword 'External'.

If this keyword is found, then we just try to find if the volume
exists, continue with using the volume as PV.

Passing gluster options to mount point is not yet complete, and
would be completed later depending on the need.