kadalu with external gluster cluster (#88)
* Enable External StorageClass through operator

* Provide another way of setting storage config

* Use external glusterfs volume as host volume of kadalu

The logic is simple. External volume details are passed as part of
storage-class itself. Instead of applying filters to find the hosting
volume, if the host-vol type passed is 'External' in storageclass,
we use the same volume for creating subvols.

Ideally, we could have better control if PVC request can have these
details, but the dynamic options are not allowed with PVC, but are
part of StorageClass.

But this is good enought to get the external storage in action for
kadalu users.

Check `examples/sample-external-kadalu-storage.yaml` for getting the
idea.

Updates #36