cli/kubectl_kadalu: add 'install' subcommand
With `install` subcommand, this CLI package now becomes
one tool to install and add storage for kadalu operator.

Invoking it is as simple as `kubectl kadalu install` and
it takes optional argument of `--version`, which by
default takes latest yaml. It also takes an argument for
letting the tool know deployment-type `--type`, which takes
argument of 'openshift' or 'kubernetes' as the option.

Fixes: #114