Quotad: Fix for 64-bit inode XFS (#334)
kadalu-quotad uses inode numbers as project numbers when setting up
quotas on XFS. Project numbers are always 32bit values, but when using
large disks XFS will generate 64 bit inode numbers.

This change modulos the inode number by 2^32 to get the inode number
in the correct range for project numbers. It does nothing good if this
results in a clash with another project number. As this seems to be
quite unlikely it feels good enough.

Signed-off-by: Alan Clucas <alan.clucas@worldprogramming.com>