Automatic detection of bricks for quotad (#337)
I'm lazy and can't be bothered with setting up a json file or
environment variable for my bricks.

By setting BRICK_NAME=AUTO this change causes quotad to query gluster
for all volumes and all bricks on those volumes, to discover any that
are on this machine.

Then we check for the 'kadalu' signature on those bricks of an info
and subvol directory in the root. If that is the case, crawl those
bricks.

The search for bricks process is time consuming, and so we only perform
it once every 30 crawls (about once a minute). This is not
configurable, just as the quota check period isn't.

Signed-off-by: Alan Clucas <alan.clucas@worldprogramming.com>