CI: tests - remove Replica2 from it (#504)
Also increase the expected_pass for CSI-sanity to 20 (from 7).

Removing replica2 from the setup is done as there were some backtraces seen with glusterfs logs in CI runs. As its not a focus in kadalu repo's files, will disable the test for now.  Will enable the tests after fixing the issues.

Signed-off-by: Amar Tumballi <amar@kadalu.io>