on-release: push operator image as last one (#605)
Currently the release script is taking ~2hrs to complete, which
means the CLI would be ready to download before the images. As
our upgrade procedure is all about restarting the operator image
with the latest version, the operator image should be the last
one to be pushed to repository, to prevent any attempt of upgrades
within that time. This also prevents any 'data' layer access issues
during upgrade.

Updates: https://github.com/kadalu/kadalu/issues/577#issuecomment-887730673

Signed-off-by: Amar Tumballi <amar@kadalu.io>