helm: Do not attempt to create the namespace (#909)
* helm: do not attempt to create the namespace

Currently, the instructions in README file advise that Helm chart has been
split into two subcharts. It then asks to deploy the Operator subchart first,
followed by the csi-nodeplugin chart. Unfortunately, that second step always
fails because, if the instructions are to be followed, the `kadalu` namespace
is implicitly created in the first step when the operator is deployed. Helm
refuses to install `cis-nodeplugin` subchart as the namespace already exists
and it is not "owned" by the `csi-nodeplugin` subchart.

IMO it's best to leave the namespace management to the user.

Signed-off-by: Marek Skrobacki <skrobul@skrobul.com>

* helm: update CHANGELOG for namespace fix

Signed-off-by: Marek Skrobacki <skrobul@skrobul.com>

Signed-off-by: Marek Skrobacki <skrobul@skrobul.com>